requirements.txt
fastapi
pydantic
python-dotenv
pymongo
uvicorn
python-multipart
azure-storage-blob
pytest
httpx
